Pull out parts into a GeoSearchInputBox component
